Day - 164 Q3

+++ B_Offshores.cpp
@@ -9,29 +9,29 @@ int main ()
     ios::sync_with_stdio(false);
     cin.tie(nullptr);
     
-    int t;
+    ll t;
     cin>>t;
 
     while(t--)
     {
-        int n,tr,cr;
+        ll n,tr,cr;
         cin>>n>>tr>>cr;
-        vector<int>v(n);
+        vector<ll>v(n);
 
-        int tt = 0;
-        for(int i=0; i<n; i++)
+        ll tt = 0;
+        for(ll i=0; i<n; i++)
         {
             cin>>v[i];
             tt += (v[i] / tr);
         }
-        int am = tt * cr;
+        ll am = tt * cr;
 
-        int rs = INT_MIN;
-        for(int x : v)
+        ll rs = LLONG_MIN;
+        for(ll x : v)
         {
-            int c = x / tr;
-            int cc = c * cr;
-            int tp = am - cc;
+            ll c = x / tr;
+            ll cc = c * cr;
+            ll tp = am - cc;
             rs = max(rs, tp+x);
         }
         cout<<rs<<'\n';
